Move Bronze table_names
