--- source/core/src/core/bronze/domain/constants/__init__.py
@@ -1,3 +1,0 @@
-from core.bronze.domain.constants.table_names import TableNames as BronzeTableNames
-
-__all__ = ["BronzeTableNames"]
