+++ source/core/src/core/bronze/infrastructure/config/__init__.py
@@ -1,0 +1,3 @@
+from core.bronze.infrastructure.config.table_names import TableNames as BronzeTableNames
+
+__all__ = ["BronzeTableNames"]
